chore: organized highlight data fetch functions
