+++ services/syntax-highlight/manage-highlight-text-on-KV.ts
@@ -1,6 +1,6 @@
 //? Import the error type for whenever there is a failed attempt to save to disk
-import { deleteFromKv } from "../../data/deleteFromKv.ts";
-import { saveToKv } from "../../data/saveToKv.ts";
+import { deleteHighlightFromKv } from "../../data/syntax-highlight/deleteHighlightFromKv.ts";
+import { saveHighlightToKv } from "../../data/syntax-highlight/saveHighlightToKv.ts";
 import FetchDataError from "../../types/error/FetchDataError.ts";
 
 //? Function to save text to disk so it can be read by another route
@@ -13,7 +13,7 @@ export async function manageHighlightTextOnKv(text: string, now: number) {
 
   //? Attempt to save a file with our data to disk
   try {
-    await saveToKv(["highlight", now.toString()], dataToSave);
+    await saveHighlightToKv(["highlight", now.toString()], dataToSave);
   } catch (error) {
     //? If the save fails, throw an error to load the error page
     throw new FetchDataError("Failed to save file to KV!", error);
@@ -23,7 +23,7 @@ export async function manageHighlightTextOnKv(text: string, now: number) {
   //? from disk after 'fileExpiration' passed
   try {
     setTimeout(() => {
-      deleteFromKv(["highlight", now.toString()]);
+      deleteHighlightFromKv(["highlight", now.toString()]);
     }, ONE_HOUR_IN_MS);
   } catch (error) {
     //? If the deletion fails, there is no need to throw an error because it wouldn't
